Update Profile functions.js
Fix PATCH `/profile` by checking if the socialMediaExtra variable has been initialized to a value other than null. Although proposed changes suggest that we stick only with validating the data agains the existing schema of updateProfile, the if statements in the code are necessary to be able to perform checks past the successfull checks of the schema validation performed in the first stage of the request handling.

I cannot successfully back any of the above statements as I haven't been able to test out the code. As such I really suggest you @apavlidi try it out and then we can make the appropriate changes.